Re-check fastpoint after reloading symbols.
Check fast tracepoints after symbols have been re-loaded.
For example, a pending tracepoint just becomes available after
a new shared object being loaded.  We didn't check it before,
because we have no idea where it is.

If the target rejects the tracepoint, an error is throw in
check_fast_tracepoint_sals, and we will disable the breakpoint.

The checking is deliberately put after the loop for adding location
to breakpoint, so users can still check the address for the tracepoint
with `info trace'.  Otherwise, it will simply show <PENDING>

(gdb) info trace
Num     Type            Disp Enb Address            What
1       fast tracepoint keep n   0x00003fffb7f507dc <pendfunc+20>
                                 ^^^^^^^^^^^^^^^^^^

This is a separate path and goes before "Tracepoint for ppc64" patch.
Is this ok?

Thanks,
Wei-cheng

--

gdb/ChangeLog

2015-08-23  Wei-cheng Wang  <cole945@gmail.com>

    * breakpoint.c (update_breakpoint_locations): Check
	fast tracepoints after reloading symbols.