Remove gdbarch_remote_breakpoint_from_pc hook.
This hook is used to get the 'kind' used in Z0/Z1 packet based on the
given pc content. In current implementation, we always set kind to 2.

When this hook is not defined, GDB will use
default_remote_breakpoint_from_pc, which delegates its work to
gdbarch_breakpoint_from_pc. So we will get the same result.

BTW, the 'kind' is not used in OpenOCD, which currently decides the
length of break insn based on the insn to be replaced.