Escape from scanning loop when jump/branch insns are met.
When optimization is on, compiler may generate code which has branch
insns before prologue like this:

   lwi333 $r3,[$r2+#0x8]
   beqz $r3,0x506356
   smw.adm $r6,[$sp],$r14,#0xa    ! {$r6~$r14, $fp, $lp
   mov55 $r10,$r0

For using jump/branch insns as terminating indication for the end of
scanning, variable `in_prologue_bb' is used to specify current
scanning status. The variable is set when those insns that only
appear in prologue is met. We know that the scanning precedure may be
ended when another jump/branch insns are met, because no jump/branch
insns are allowed in prologue basic block.