Remove CPC in nds32_analyze_prologue.
During skipping prologue, the scanning precedure is stopped once the
insn is not recognized, and the corresponding PC is used as the end
of prologue.

CPC is the pc whose insn is to adjust stack, and it is used as the end
of prologue only when all insns are recognized before the set limit pc.

Removing CPC mechanism, because:
1. Only compiler-generated codes are focused on.
2. For the compiler-generted codes, there must be some unrecognized
insns before the set limit pc.
3. No other architectures use similar mechanism.