Change type of sp_offset/fp_offset to unsigned.
LONGEST is typedef of long long.
CORE_ADDR is typedef of unsigned long.

saved_regs[] is used to record the memory address where registers are
saved, but, during prologue analysis, this array is also used to record
the offset from the prev_sp temporarily.

Change type to avoid the conversion between signed and unsigned type.